+++ src/i18n/translations/index.ts
@@ -1,5 +1,8 @@
+import de from './de.json';
 import en from './en.json';
 import es from './es.json';
+import fr from './fr.json';
+import zh from './zh.json';
 
 /**
  * Translation Resources
@@ -17,5 +20,26 @@ export const resources = {
    */
   es: {
     translation: es,
+  },
+
+  /**
+   * French
+   */
+  fr: {
+    translation: fr,
+  },
+
+  /**
+   * German
+   */
+  de: {
+    translation: de,
+  },
+
+  /**
+   * Chinese
+   */
+  zh: {
+    translation: zh,
   },
 } as const;