+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Features / Enhancements
 
 - Add yearly view (#134)
-- Add internationalization(Spanish, French, German, and Chinese) i18n (#135)
+- Add internationalization(Spanish, French, German, and Chinese) i18n (#135, #138)
 - Add default view option (#137)
 
 ## 2.2.0 (2023-10-06)